requirements.txt
aiobotocore==2.11.2
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
anyio==4.3.0
attrs==23.2.0
boto3==1.34.33
botocore==1.34.48
certifi==2024.2.2
et-xmlfile==1.1.0
frozenlist==1.4.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
jmespath==1.0.1
multidict==6.0.5
notion-client==2.2.1
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
s3fs==2024.2.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
tzdata==2024.1
urllib3==2.0.7
wrapt==1.16.0
yarl==1.9.4
